pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "werkzeug>=3.1.4",
]
